pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chromascope"
version = "0.1.0"
description = "Audio analysis engine for reactive generative art visualizations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Developer"}
]
keywords = ["audio", "analysis", "visualization", "music", "librosa"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "soundfile>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
experiment = [
    "Pillow>=10.0.0",
    "noise>=1.2.2",
    "numba>=0.57.0",
    "pygame>=2.1.0",
]

[project.scripts]
chromascope = "chromascope.cli:main"
chromascope-fractal = "chromascope.experiment.cli:main"
chromascope-solar = "chromascope.experiment.solar_cli:main"
chromascope-decay = "chromascope.experiment.decay_cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/chromascope --cov-report=term-missing"
